move distribution wrapper to separate file
